Extract update my cart/customer into its own mixin
