+++ src/components/UpdatableCartInfo.gql
@@ -1,3 +1,5 @@
+#import "./DisplayableMoney.gql"
+
 fragment UpdatableCartInfo on Cart {
   id
   version
@@ -36,6 +38,7 @@ fragment UpdatableCartInfo on Cart {
   }
 }
 
+# import does not work https://github.com/apollographql/graphql-tag/issues/159
 fragment DisplayableMoney on Money {
   centAmount
   currencyCode
