+++ src/components/UpdatableCustomerInfo.gql
@@ -1,0 +1,7 @@
+fragment UpdatableCustomerInfo on Customer {
+  id
+  version
+  email
+  firstName
+  lastName
+}
